backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Frontier AI Solutions Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.120.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
    "websockets>=12.0",
    "azure-cognitiveservices-speech>=1.38.0",
    "openai>=1.0.0",
    "azure-identity>=1.15.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.14.2",
    "requests>=2.32.5",
    "lxml>=6.0.2",
    "redis>=7.0.1",
]
